Refactored Structure and Added Server Template
I created seperate directories that contain the algorithm and the threaded server surrounding the algorithm as well. Additionally, I created a server template that accepts connections on port 3002 and spins up a new process for each connection and is able to accept a series of messages incoming from the specific client.

To-do: Create a data input structure that matches a series of messages over the socket with data input for courses.