Can now transmit/receive a complete schedule
The server is now able to handle an incoming schedule where all names are sent as a new-line terminate string and each time and course duration is transmitted as a new-line terminated string parsable to an integer.

To-do: Hook the server up to an instantiated algorithm and transmit the result back over the socket.