Added Seed and Crossover
Work Done:

Added a new schedule class that contains a list of sections. Can be instantiated simply by giving it a section array OR by giving it a hashmap between strings and sections and a boolean double array of sections.

Added a seed method that generates a starting schedule seed that could contain conflicts but is entirely valid.

Added a crossover method that randomly crosses schedules, represented as boolean arrays, over then generates a new schedule with the results.