Added Course Name Conflicts Checker
Added a method that checks for any course name conflicts then generates a complete fitness score based on the combination of the factors.

This algorithm currently doesn't check direction of convergence, but doesn't need to because the existence of a valid schedule is binary -- there either exists or does not exist a valid schedule.