Completed the NetWork Handler
The network handler class is a wrapper that makes a Java socket compliant with a W3C Websocket.

First, it completes the websocket handshake that allows it to communicate between the client and the server.

After initialization, the client may send messages which will be received upon a query by the server. There are no automatic callbacks, and the request for a new message will halt the program until a message is received.

The server may use the sendMessage() method to send any string message.