Added the create/join group functionality
Users can now create a group and receive a custom URL for inviting others to join their group.
Users can join a group through using the URL.
Currently, the join page does not check if a user is logged in or check to see if the user is in the group.

Need to find a way to get the group name before using the joingroup function.