Encapsulation and Calculation
Abstracted the calculation of both optional and required scores (which both occur at the same time) to a single class that can be invoked and contains information about the convergence of all parameters.

Created a new test to measure the convergence of the scores on a sliding window of 100. The Root Mean Squared test for the difference between a given value and the mean value on the window tends to 0 as the convergence similarly tends to 0.